LayoutTests/media/track/track-webvtt-tc002-bom-expected.txt

Tests that the parser properly ignores a UTF-8 BOM character at the beginning of a file and all other cues are properly parsed.


*** Testing text track 0
EXPECTED (cues.length == '2') OK
EXPECTED (cues[i].id == '1') OK
EXPECTED (cues[i].startTime == '0') OK
EXPECTED (cues[i].endTime == '30.5') OK
EXPECTED (cues[i].getCueAsSource() == 'Bear is Coming!!!!!') OK
EXPECTED (cues[i].id == '2') OK
EXPECTED (cues[i].startTime == '31') OK
EXPECTED (cues[i].endTime == '1200.5') OK
EXPECTED (cues[i].getCueAsSource() == 'I said Bear is coming!!!!') OK
END OF TEST
